Changed data type from Long to Double for CurrentTime measurement
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio.ProbeSpecification/trunk/de.uka.ipd.sdq.probespec.framework@8459 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ src/de/uka/ipd/sdq/probespec/framework/probes/ATakeCurrentTimeStrategy.java
@@ -16,7 +16,7 @@ public abstract class ATakeCurrentTimeStrategy implements IProbeStrategy {
 	/**
 	 * {@inheritDoc}
 	 */
-	public abstract ProbeSample<Long, Duration> takeSample(String probeId,
+	public abstract ProbeSample<Double, Duration> takeSample(String probeId,
 			Object... o);
 
 }
